fixed a bug where you couldn't save point value

+++ Tamagotchi/Tamagotchi_Library/GameManagement/SaveManager.cs
@@ -58,7 +58,7 @@ namespace Tamagotchi_Library.GameManagement
         public void updateData(IFeline toSave)
         {
             felineData = toSave.getFelineInfo();
-            age = int.Parse(felineData[1]);
+            age = float.Parse(felineData[1]);
             hunger = int.Parse(felineData[2]);
             thirst = int.Parse(felineData[3]);
             happiness = int.Parse(felineData[4]);
